refactor: moved chats logic in new <Filter>
& add new filtersSlice in redux store
