+++ src/components/Chat/Chat.module.scss
@@ -18,148 +18,6 @@
   }
 }
 
-.chat_icons {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  gap: 15px;
 
-  border-right: 1px solid $grey;
-  border-left: 1px solid $grey;
-  padding: 0 5px;
+/* delete styles below */
 
-  span {
-    font-size: 16px;
-  }
-}
-
-.date_container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  input {
-    width: 104px;
-    border-radius: 5px;
-    padding: 5px 0;
-    font-family: inherit;
-    font-size: 13px;
-    background-color: $darkreader-selection-text;
-  }
-}
-
-.filter {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: transparent;
-  border: 0;
-  width: 38px;
-  height: 38px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.5s;
-
-  img {
-    width: 35px;
-    height: 35px;
-    background-size: contain;
-  }
-
-  &:hover {
-    background-color: $highlight-hover-color;
-  }
-}
-
-.filter2 {
-  img {
-    width: 25px;
-    height: 25px;
-  }
-}
-
-.close_filter {
-  position: absolute;
-  top: -5px;
-  right: 14px;
-  opacity: 0.6;
-  cursor: pointer;
-  transition: opacity ease 0.5s;
-  width: 5px;
-
-  &:hover {
-    opacity: 1;
-  }
-}
-
-.close_filter_2 {
-  top: 0px;
-  right: 13px;
-}
-
-.close_filter_3 {
-  top: 0px;
-  right: 13px;
-}
-
-.close_filter::before,
-.close_filter::after {
-  content: '';
-  position: absolute;
-  top: 10px;
-  display: block;
-  width: 15px;
-  height: 3px;
-  background: #d8d4cf;
-  border-radius: 100px;
-}
-
-.close_filter::before {
-  transform: rotate(45deg);
-}
-
-.close_filter::after {
-  transform: rotate(-45deg);
-}
-
-.filters_container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 10px;
-  position: relative;
-}
-
-.user_filter_container {
-  display: flex;
-  gap: 10px;
-  padding: 13px 17px 13px 0px;
-  position: relative;
-
-  .user_filter {
-    display: flex;
-    align-items: center;
-    gap: 5px;
-    background-color: $darkreader-bg2;
-    border-radius: 20px;
-    padding: 5px;
-    transition: background-color ease 0.5s;
-    cursor: pointer;
-
-    span {
-      font-size: 14px;
-    }
-
-    img {
-      background-color: white;
-      height: 30px;
-      width: 30px;
-      border-radius: 50%;
-      object-fit: cover;
-    }
-
-    &:hover {
-      background-color: $highlight-hover-color;
-    }
-  }
-}
