+++ src/store/index.ts
@@ -2,6 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import authReducer from './authSlice';
 import chatReducer from './chatsSlice';
 import editReducer from './editMessageSlice';
+import filterReducer from './filtersSlice';
 
 export type IRootState = ReturnType<typeof store.getState>;
 
@@ -10,5 +11,6 @@ export const store = configureStore({
     auth: authReducer,
     chat: chatReducer,
     edit: editReducer,
+    filters: filterReducer,
   },
 });
